backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
openai>=1.14.0
python-dotenv>=1.0.0
